Adaptations to pass scikit-hep repo review (#128)
This patches adds a lot more checks to ruff and pre-commit.

Major effort was put into typing everything, which was particularly
challenging in this project, that does a lot of dynamic compilation.

A benefit of this work is the `_jit` decorator and its variants could
now generate suitable signatures for the jitted functions based on the
arguments of the original function, so that the number of arguments
doesn't have to be given explicitly. I leave that for a follow-up work.